[offload] Decouple PluginManager from the OpenMP offload policy handling
This patch refactors libomptarget to decouple OpenMP specific handling from
PluginManager. This is to prepare PluginManager to be split off into a separate
library to be used for both the OpenMP and OpenACC runtime.
